refactor login strategies into own service
